requirements.txt
fastapi
uvicorn[standard]
pytesseract
numpy
pandas
pysqlite3
requests
scikit-learn
Pillow
matplotlib